+++ CHANGELOG.md
@@ -32,6 +32,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids incorrect exception order
 - Enforces tuples usage with frozenset constructor
 - Changes how `WPS444` works, now we use stricter logic for `while` and `assert`
+- Forbids to use `yield from` with incorrect types
 
 ### Bugfixes